All plugin .pro files now fix library dependencies to address QT 5.2 crash when plugins were loaded
Pebble plugins were referencing the installed QT deployment, while Pebble.app was referencing the QT libraries added by macdeployqt tool.  The result was that when a plugin executed code that relied on globals like qApp (needed to create new UI elements) it would crash with an error "QWidget: Must construct a QApplication before a QWidget".  This checkin manually fixes each plugin so it references the Frameworks directory in pebble.app