Switched from SDR::Factory to plugin load
Internal SDR objects are now created when pseudo-plugins are created in plugins.cpp.  These are assigned to the deviceInterface member, just as a regular plugin would be when loaded.  So we now have permanent objects for every SDR and no longer have to create and destroy them.  The active device deviceInterface is kept in global.sdr, so it is always available.  Internally sir variable is still used everywhere, it's just gets its value from global->sdr when the receiver is turned on