FileSDRDevice plugin works!
Switched to using SDR as 'home' class for all internal and plugin devices.  SDR now has a pointer to plugin if that's the active device.  If so, it delegates interface calls through that pointer to the plugin.  If *plugin is NULL, then we are using internal devices and calls are routed through inheritance.  All these internal devices will eventually disappear and we'll be left with SDR as a delegator.