Major update to the way Producer/Consumer is implemented
No longer uses subclassed QThread, used Worker objects that are moved to an instance of QThread.  This is the new recommendation from QT group for using QThreads

Added ability to only start producer, consumer or both.

Added ability to specify timeout value for acquiring a buffer.  This fixed a problem with deadlock in the rtl2832 plug in.  If we don't get a buffer within the timeframe (ms) we return false so the code using it can handle.

Misc cleanup