Major change in packaging.  All dependent dylibs are now found in lib directory
We need the dylib files to be available to other programs using PebbleLib and plugins, so the old approach of copying them into the app via macdeployqt doesn't work.  Instead we use install_tool to manual fix up all links to use @rpath.  rpath is then set in each quake file that uses these plugins or dylibs and points to the location of the lib directory, relative to the application