Bug: QAudioOut buffer overflow
QAudioOut seems to only be able to buffer 32767 bytes.  This makes it easy to overflow, resulting in lost data.  We do what we can here, but the right fix is to create our own IO class that manages a reasonable buffer strategy.
Note:PortAudio does not have this problem and deals with it internally