Support higher sampling rates
Allow producer/consumer timer interval to be set to zero, which is the fastest QTimer allows.  In this mode, the worker functions will be called whenever there are no events to be processed.  Typically this is the case when SR is greater than 2048sps.  Note high CPU when interval is zero