Fix audio distortion
We were driving samples above 1.0 and below -1.0 by just multiplying by the gain factor.  As soon as we fixed clipping, this became apparent.

For QAudio, we added the setVolume() call and pass it a value from 0.0.
For Port audio, we still scale from 0 to 1

Adjustments to NFM and WFM gain to match audio output