Clean up AGC ui changes
We were getting duplicate and confusing signals when receiverWidget changed agc mode.  It was triggering a signal for the slider reset and then another signal for the mode change.
Now receiverWidget is solely responsible for setting slider.  agcModeChanged signal now has the new threshold value that the slider is set to.