Yet another tweak to k, support for neg freq
More testing revealed that adding .5 to k was putting us off center in the bin, so reverted.
Support for negative freq to allow proper use with RF.  So morse modem tone/freq can be -1000 for cwl or +1000 for cwu.
Tests perfectly.

Changed attack/decay variables for a more reasonable decay.  Seems to perform better in noisy conditions